+++ docs/CHANGELOG.md
@@ -47,6 +47,7 @@ RELEASE 1.5.5-20140113
 - Fix broken Contact's template (#2)
 - Fix printing all records from RecordBrowser
 - Fix watchdog email notifications (#3)
+- Fix TCPDF top margin when logo is set but it's hidden
 - Update translations
 
 RELEASE 1.5.4-rev11060 (20131015)